+++ locales/index.js
@@ -1,10 +1,12 @@
-const { LOCALES_EQUIVALENTS_MAP, LANGUAGES } = require('../scripts/locales/consts');
+const { getEquivalent } = require('../scripts/locales/helpers');
+
+const { LANGUAGES, LOCALE_DATA_FILENAME } = require('../scripts/locales/constants');
 
 const locales = Object.keys(LANGUAGES)
     .reduce((acc, language) => {
-        const resultLocale = LOCALES_EQUIVALENTS_MAP[language] || language;
+        const resultLocale = getEquivalent(language);
         // eslint-disable-next-line global-require,import/no-dynamic-require
-        const dictionary = require(`./${resultLocale}/messages.json`);
+        const dictionary = require(`./${resultLocale}/${LOCALE_DATA_FILENAME}`);
         acc[resultLocale] = dictionary;
         return acc;
     }, {});