+++ scripts/locales/constants.js
@@ -9,6 +9,8 @@ const [{
     localizable_files: LOCALIZABLE_FILES,
 }] = require('../../.twosky.json');
 
+const LOCALES_DIR = '../../locales';
+
 /**
  * Users locale may be defined with only two chars (language code)
  * Here we provide a map of equivalent translation for such locales
@@ -18,10 +20,41 @@ const LOCALES_EQUIVALENTS_MAP = {
     'zh-CN': 'zh',
 };
 
+const LOCALES = Object.keys(LANGUAGES);
+
+const REQUIRED_LOCALES = [
+    'de',
+    'es',
+    'fr',
+    'it',
+    'ja',
+    'ko',
+    'pt',
+    'pt-PT',
+    'ru',
+    'zh',
+    'zh-TW',
+];
+
+/**
+ * Threshold percentage for validation of required locales.
+ */
+const THRESHOLD_PERCENTAGE = 100;
+
+/**
+ * Name of the locale data file.
+ */
+const LOCALE_DATA_FILENAME = 'messages.json';
+
 module.exports = {
     LOCALES_EQUIVALENTS_MAP,
     BASE_LOCALE,
     PROJECT_ID,
     LANGUAGES,
     LOCALIZABLE_FILES,
+    LOCALES,
+    REQUIRED_LOCALES,
+    THRESHOLD_PERCENTAGE,
+    LOCALES_DIR,
+    LOCALE_DATA_FILENAME,
 };