+++ scripts/locales/locales.js
@@ -1,36 +1,36 @@
-/* eslint-disable no-console */
 const fs = require('fs-extra');
 const path = require('path');
 const axios = require('axios');
+
 const {
     BASE_LOCALE,
     PROJECT_ID,
-    LANGUAGES,
     LOCALIZABLE_FILES,
-    LOCALES_EQUIVALENTS_MAP,
-} = require('./consts');
+    LOCALES,
+    LOCALES_DIR,
+} = require('./constants');
+const { checkTranslations } = require('./validate');
+const { getEquivalent } = require('./helpers');
+const { cliLog } = require('./cli-log');
 
 // URLs
 const BASE_URL = 'https://twosky.int.agrd.dev/api/v1';
 const BASE_DOWNLOAD_URL = `${BASE_URL}/download`;
 const BASE_UPLOAD_URL = `${BASE_URL}/upload`;
-// Directory where locales should be stored
-const LOCALES_DIR = './locales';
-const CROWDIN_PROJECT = PROJECT_ID;
-const LOCALES = Object.keys(LANGUAGES);
+
 const CROWDIN_FILES = LOCALIZABLE_FILES
     .map((pathToFile) => pathToFile.split('/')
         .pop());
 
 /**
- * Build query string for downloading tranlations
+ * Build query string for downloading translations
  * @param {string} lang locale code
  * @param {string} file crowdin file name
  */
 const getQueryString = (lang, file) => {
     let res = '?format=json';
     res += `&language=${lang}`;
-    res += `&project=${CROWDIN_PROJECT}`;
+    res += `&project=${PROJECT_ID}`;
     res += `&filename=${file}`;
     return res;
 };
@@ -63,22 +63,16 @@ const removeEmptyStrings = (data) => {
 };
 
 /**
- * Returns equivalent of specified locale code
- * @param {string} locale locale
- */
-const getEquivalent = (locale) => LOCALES_EQUIVALENTS_MAP[locale] || locale;
-
-/**
- * Build form data for uploading tranlation
+ * Build form data for uploading translation
  * @param {string} file crowdin file name
  */
 const getFormData = (file) => {
-    const pathToBaseFile = path.resolve(LOCALES_DIR, BASE_LOCALE, file);
+    const pathToBaseFile = path.resolve(__dirname, LOCALES_DIR, BASE_LOCALE, file);
     const body = new FormData();
 
     body.append('format', 'json');
     body.append('language', BASE_LOCALE);
-    body.append('project', CROWDIN_PROJECT);
+    body.append('project', PROJECT_ID);
     body.append('filename', `${file}`);
     body.append('file', fs.createReadStream(pathToBaseFile));
     return body;
@@ -110,16 +104,17 @@ async function download() {
         for (const file of CROWDIN_FILES) {
             const url = getDownloadURL(lang, file);
             try {
-                console.log('Downloading:', url);
+                cliLog.info(`Downloading: ${url}...`);
                 // eslint-disable-next-line no-await-in-loop
                 const { data } = await axios.get(url);
                 const resultLocale = getEquivalent(lang);
-                const filePath = path.resolve(LOCALES_DIR, resultLocale, `${file}`);
+                const filePath = path.resolve(__dirname, LOCALES_DIR, resultLocale, `${file}`);
                 const formatted = removeEmptyStrings(data);
                 saveFile(filePath, formatted);
+                cliLog.success(`Downloaded ${file} for ${resultLocale}`);
             } catch (e) {
-                console.log('Error on downloading url:', url);
-                console.log(e.message);
+                cliLog.error(`Error on downloading url: ${url}`);
+                cliLog.error(e.message);
             }
         }
     }
@@ -128,7 +123,7 @@ async function download() {
 /**
  * Entry point for uploading translations
  */
-function upload() {
+async function upload() {
     CROWDIN_FILES.forEach(async (filename) => {
         try {
             const formData = getFormData(filename);
@@ -136,23 +131,49 @@ function upload() {
                 contentType: 'multipart/form-data',
                 headers: formData.getHeaders(),
             });
-            console.log('Successfully uploaded');
+            cliLog.success(`Successfully uploaded ${filename}`);
         } catch (e) {
-            console.log(e);
+            cliLog.error(e.message);
         }
     });
 }
 
-/**
- * You need set environment variable LOCALES=DOWNLOAD|UPLOAD when run the script
- */
-if (process.env.LOCALES === 'DOWNLOAD') {
-    download().catch((e) => {
-        console.log(e);
-        process.exit(1);
-    });
-} else if (process.env.LOCALES === 'UPLOAD') {
-    upload();
-} else {
-    console.log('Option DOWNLOAD/UPLOAD locales is not set');
+async function validateRequired() {
+    await checkTranslations(LOCALES, { isMinimum: true });
 }
+
+async function validateAll() {
+    await checkTranslations(LOCALES);
+}
+
+/**
+ * You need set environment variable LOCALES when run the script. Available options:
+ * - DOWNLOAD
+ * - VALIDATE-ALL
+ * - VALIDATE-MIN
+ * - UPLOAD
+ */
+const main = async () => {
+    if (process.env.LOCALES === 'DOWNLOAD') {
+        try {
+            await download();
+        } catch (e) {
+            cliLog.error(e.message);
+            process.exit(1);
+        }
+        await validateRequired();
+    } else if (process.env.LOCALES === 'VALIDATE-ALL') {
+        await validateAll();
+    } else if (process.env.LOCALES === 'VALIDATE-MIN') {
+        await validateRequired();
+    } else if (process.env.LOCALES === 'UPLOAD') {
+        await upload();
+    } else {
+        cliLog.error('No valid LOCALES environment variable set');
+        process.exit(1);
+    }
+};
+
+(async () => {
+    await main();
+})();