+++ CHANGELOG.md
@@ -7,6 +7,22 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [4.3.74] - 2025-06-30
+
+### Added
+
+- Support of Bulgarian, Catalan, Macedonian and Thai locales.
+
+### Changed
+
+- Updated UI of element selector menu to match with other projects.
+
+### Fixed
+
+- Issue with slider dragging does not work properly on touch devices.
+
+[4.3.74]: https://github.com/AdguardTeam/AdguardAssistant/compare/v4.3.70...v4.3.74
+
 ## [4.3.70] - 2023-06-01
 
 ### Added